update: fixing modal
