+++ frontend/src/components/modal/Modal.tsx
@@ -1,72 +1,24 @@
-import { useDispatch, useSelector } from "react-redux";
-import { toggleModal } from "../../utils/redux/reducers/modalSlice";
-import { useAppSelector, useAppDispatch } from "../../hooks/ReduxHooks";
-import { RootState } from "../../utils/redux/store";
+import React from "react";
+import ReactDOM from "react-dom";
 import "./Modal.css";
 
-function Modal() {
-  const isOpen = useAppSelector((state: RootState) => state.modal.isOpen);
-  const dispatch = useAppDispatch();
-  const toggle = () => {
-    dispatch(toggleModal());
-  };
-  return (
-    <>
-      {isOpen && (
-        <div className="modal-overlay">
-          <div className="modal">
-            <div id="form">
-              <h2>Nova Transação</h2>
-              <form action="">
-                <div className="input-group">
-                  <label className="sr-only" htmlFor="description">
-                    Descrição
-                  </label>
-                  <input
-                    type="text"
-                    id="description"
-                    name="description"
-                    placeholder="Descrição"
-                  />
-                </div>
-
-                <div className="input-group">
-                  <label className="sr-only" htmlFor="amount">
-                    Valor
-                  </label>
-                  <input
-                    type="number"
-                    step="0.01"
-                    id="amount"
-                    name="amount"
-                    placeholder="0,00"
-                  />
-                  <small className="help">
-                    Use o sinal - (negativo) para despesas e , (vírgula) para
-                    casas decimais
-                  </small>
-                </div>
-
-                <div className="input-group">
-                  <label className="sr-only" htmlFor="date">
-                    Data
-                  </label>
-                  <input type="date" id="date" name="date" />
-                </div>
-
-                <div className="input-group actions">
-                  <a onClick={toggle} href="#" className="button cancel">
-                    Cancelar
-                  </a>
-                  <button>Salvar</button>
-                </div>
-              </form>
-            </div>
-          </div>
-        </div>
-      )}
-    </>
-  );
+interface ModalProps {
+  children: React.ReactNode;
+  isOpen: boolean;
 }
+
+const Modal: React.FC<ModalProps> = ({children, isOpen}) => {
+if (isOpen == false) {
+  return null;
+}
+  return ReactDOM.createPortal(
+    <div className="modal-overlay active">
+      <div className="modal">
+        {children}
+      </div>
+    </div>,
+    document.getElementById("modal-root") as HTMLElement
+  );
+};
 
 export default Modal;
