+++ frontend/src/features/expenses/Expenses.css
@@ -55,6 +55,7 @@ a {
 
 a:hover {
   color: var(--light-green);
+  cursor: pointer;
 }
 
 button {
@@ -209,43 +210,6 @@ td.income {
 
 td.expense {
   color: #e92929;
-}
-
-/* Form ===================== */
-#form {
-  max-width: 500px;
-}
-
-#form h2 {
-  margin-top: 0;
-}
-
-input {
-  border: none;
-  border-radius: 0.2rem;
-
-  padding: 0.8rem;
-
-  width: 100%;
-}
-
-.input-group {
-  margin-top: 0.8rem;
-}
-
-.input-group .help {
-  opacity: 0.4;
-}
-
-.input-group.actions {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.input-group.actions .button,
-.input-group.actions button {
-  width: 48%;
 }
 
 /* Footer ===================== */
